Add `From` conversion for `T: WndListener` etc.
`set_listener` and similar methods now accept concrete types, which
are boxed automatically. Reduces the amount of boilerplate code in most
cases.

There are two shortcomings with this change: (a) The methods are now
generic and might lead to an increase in compilation time and compiled
code size because of monomorphization. (b) The parameter types are now
only bound by `impl Into<_>`, weaker than the concrete type `Box<_>`.