Add `tcw3::ui::images`
This new module provides an application-global image manager that
automatically rasterizes and caches images for requested DPI scale
values. It’s intended to be used to facilitate theming and icon
rendering.

Components for displaying images using this module are to be added.
More `Img` types (e.g., for basic figures and StVG images) are to be
added.