Add a new UI styling system
This commit adds a new UI styling system used to decorate UI elements
based on stylesheets. The primary goal of this addition is separation
of styling from behavioral code, and is not to allow application users
to provide their own custom themes. However, that would be possible
too if an appropriate interface was designed.