Implement `Lineset::regroup` (WIP)
This commit add a preprocessing step that subdivides LOD groups (up to
LOD level 1) until the precise indices of viewports are known.

This step is similar to the split step, but recalculates the affected
index range based on the viewport's position range as the subdivision
proceeds.

Why didn't I simply replace the split step with this? It's not simple
because the goal list must be updated as lines are resized. The merge
step would probably be more complicated when endpoints are specified by
line coordinates.