Store `TWM` in `WinitWm<TWM>`
The possession of `TWM` entails the main thread is the owner. By storing
one in `WinitWm`, the accessor of `WinitWm` can safely claim that it's
running in the main thread.

As a result, `<tcw3::pal::macos::WM as tcw3::pal::iface::WM>::
invoke_on_main_thread` no longer requires an unsafe block.