Refactor `Table`
Viewport position was moved from `Inner` to `State` and is now
encapsulated by `VpSet`.

`Table` needs to be able to handle multiple viewports. There is one
permanent viewport for the visible portion. While scrolling, another
viewport is needed to "pin" the original scroll position.

When updating the model, viewports have to be adjusted to follow the
movement of lines (we call this the application of *a displacement
policy*). Since this is done in multiple places, it's preferable to
pack them into a single `struct` so that, when we finally add support
for multiple viewports, most code is unaffected.