testing: Borrow some modules from `unix` backend
As it turned out, `cfg(feature = "name")` does not work in `Cargo.toml`.

The module organization is a little bit messy; `unix/{bitmap.rs,text.rs}`
are included in two modules: `unix` and `testing`, and for the latter
`#[path = "..."]` is used to override the standard module loading rule.
Hopefully this could be refactored in the future.