testing: Implement and use `UniqPool`
`UniqPool` is a wrapper around `iterpool::Pool` that ensures the
uniqueness of pointers, meaning every inserted elements are guaranteed
to have unique pointers, and the same pointer value is never reused for
an element inserted later.

The reason of this change is to prevent the methods (to be added in
upcoming commits) for querying window states from succeeding spuriously.