testing: Add `tcw3_testing`
This crate provides a procedural attribute macro `#[use_testing_wm]`
that wraps a function to use the testing backend. Example code:

    #[use_testing_wm]
    #[test]
    fn new_wnd(twm: &dyn TestingWm) {
        let _hwnd = twm.wm().new_wnd(Default::default());
    }

`tcw3_testing` can receive a feature flag `testing`. This serves no
purpose at the moment. One possibility is to add another attribute macro
that behaves like `#[cfg(feature = "tcw3_pal/testing")]`. A portion of
`tcw3_pal`'s testing-specific code is compiled in regardless of `testing`
feature flag so that unit tests do not break when `testing` is disabled.
This macro would let us remove that portion of code. But it would have
to be attached to each relevant item, which can be tedious, because
using an attribute macro on modules is unstable at the point of writing.

Another possibility is to control the behavior of `#[use_testing_wm]`.
For example, it could output a warning during compilation when `testing`
is disabled. Unfortunately, I'm not aware of any possible way to output
a warning from a procedural macro. It also could suppress the generation
of annotated functions to save compilation time.