Fix assertion failure in `Wnd::drop`
This was due to an incorrect assertion condition. `Wnd::drop` is called
when the ref count of `Rc<Wnd>` falls to zero. This invalidates all
`Weak<Wnd>` associated with the window, including the one in the
superview pointer of the window's content (root) view. The assertion
fired because it did not handle this case.