Turn off `clippy::float_cmp` globally
This lint is ridiculous, should be moved to `pedantic` (off by default)
and have its suggestion rewritten.

I know it's designed to catch the misuses of floating-point numbers
where they are merely approximate representation of real numbers but
they are treated as if having infinite precision. But this is a novice
mistake that can be avoided only by the shallowest level of
understanding of floating-point numbers, and must not be avoided by
blindly replacing them with approximate comparison idioms. Also, this
lint flags too many legit uses of floating-point comparison.