PAL: Define scroll events
`ScrollListener::motion` provides `velocity` through a parameter because
the receiver might not be able to estimate it reliably because of
non-real-time scheduling. Providing a timestamp isn't an option either
because `winit` doesn't support that.

This definition should work for macOS (legacy mouse + Magic Mouse) and
winit (legacy scrolling only). For Wayland, there is
`wl_pointer::axis_stop` event generated when fingers are lifted to start
inertial scrolling. Our event definition should be able to accommodate
the event (though PAL would have to synthesize inertial movements), but
I'm not sure how the event is translated by winit.