designer: Include dependencies in metadata
This change makes it possible for field definitions to refer to other
crates' component types.

Growth in the metadata size might be concerning.

Crates are identified by UUIDs and not by names, allowing two different
versions of the same crate to co-exist in a single top-level unit of
metadata (called a repository). This behavior is faithful to Rust's
type system. While it may seem confusing, it actually clear things up
a bit in light of the possibility to import crates under a different
name.