designer: Improve the checking of invalid component paths
The check whether the path is inside the current crate or not was
previously done on AST, meaning it had to handle multiple cases,
e.g., `crate::`, `crate_name::` are okay, but not `::crate` unless the
crate is named `crate`. It's now done after resolving the crate index
for more streamlined implementation.