[refactor] designer: Check and analyze `ObjInit` in an earlier stage
This prepares the code for the implementation of the simplified object
initialization syntax (based on the More Concise Object Initialization
proposal).

In the new syntax, field types are no longer specified and included in
`sem::ObjInitField`. There are some occasions where the field types are
needed, so this commit introduces a new analysis phase that resolves
field references and figures out the field types being initialized.