gtk: Do not cancel a mouse drag gesture on `leave` event
The `leave` event can be raised even if a mouse button is pressed. If
we cancel the active mouse drag gesture in the event handler, it results
in an annoying behavior.

This change causes another problem: The drag gesture is not canceled if
the user switches to another application while pressing a mouse button
down. Other GNOME applications (but not all of them) exhibit a similar
issue, so I think it's acceptable.