Clear `View::layers` on unmount
This fixes panics in `stella2` when toggling the sidebar.

`ViewListener::unmount` removes associated layers, so `View::layers`
should be cleared as well. What happens otherwise? The usual
implementation of `ViewListener::update` calls `ctx.layers()` to check
if the stored layer set is up-to-date or not. Often, it only checks the
number of elements as a shortcut, meaning it will not replace an old
layer set if it has the same number of elements as the latest one.
The old layer set may contain layer handles which are not valid anymore,
so a panic ensues.