Add `ScrollableTable::set_size_traits`
This is mainly for convenience for the writers of `.tcwdl`. The obj-init
syntax only can set the direct (not subcomponent's) fields of the
initialized object. Thus, setting the inner `Table`'s `size_traits` is
beyond the capability of the obj-init syntax and necessitates the use of
`on (init)`. This commit adds the `ScrollableTable::set_size_traits`
method that makes it available as a direct prop of `ScrollableTable`.