fix(view): raise `dispatch` when a splitter is moved
This functionality was temporarily removed in
1d00248ea5f675e396669ca81854136fefc7dbab because there was no obvious
way to subscribe to the event handler back then.

Actually, this fix does not rely on any new feature. The fix uses an
adapter component (written using `#[prototype_only]`) to convert a drag
completion event to a Designer-compatible event. This has always been
possible since when Designer was merged into `master`.