fix(gtk): call `Screen::update_wnd` on surface renewal
`tcw_wnd_widget_draw_handler`, when it detects that surface properties
(such as size) are updated, calls `WndListener::resize`. During this,
calls to `Wm::update_wnd` are ignored in  `WndListener::resize` because
we want the corresponding repainting process to complete before
returning from `tcw_wnd_widget_draw_handler`. The repainting process is
initiated by `tcw_wnd_widget_draw_handler` calling
`Compositor::update_wnd`.

Formerly, backends were not allowed to implicitly call `Wm::update_wnd`,
so the state of a layer tree was preserved until the client explicitly
called `Wm::update_wnd`. `Compositor::update_wnd` took a parameter
named `update_layers`, which essentially provides a way to suppress the
behavior of `Wm::update_wnd`. This is not relevant anymore, but whether
`Wm::update_wnd` is implicit or not has nothing to do with the problem
this commit intends to fix.

The problem happens when `Wm::update_wnd` is called inside `WndListener
::resize`. In this case, the call to `Wm::update_wnd` is ignored as
explained earlier. Nevertheless, `tcw_wnd_widget_draw_handler` later
calls `Compositor::update_wnd` with `update_layers = false`, instructing
the compositor to preserve the old state of the layer tree (this would
be okay for normal draw calls which do not involve surface renewal, but
it's not in this case). This means the call to `Wm::update_wnd` does not
actually take effect until it's called again outside `WndListener::
resize`.

This commit removes the said parameter of `Compositor::update_wnd`
(essentially locking it to `true`) to address the problem.