feat(windows): abort the process immediately on panic
This commit enables the `panic_immediate_abort` feature of `libstd`,
which changes the panic behavior to abort the process immediately
without formatting the panic message.

By default, panic messages are outputted to stderr, which is useless
for a Windows GUI application.

I plan to use the `MiniDumpWriteDump` Win32 API function to generate a
minidump for bug reporting. A minidump provides a complete backtrace as
well as the state of local variables, which may be more or less useful
than a panic message.

This change reduces the output executable size by 119KB (-21.4%). This
change only affects the binaries built using Xargo; those built by
`cargo build` as usual aren't affected.