fix(ui): remove the auto classes `HOVER` and `FOCUS` on mount
A view may be removed while having a keyboard focus. In this case, the
keyboard focus will be lost, but the view doesn't receive a focus lost
event. Later when the view is re-added to a window, the view is left
unfocused. This means the view should revert itself to an unfocused
state in the mount event handler.

The same goes for mouse hover events.