feat(utf16count): add `utf16count`
Cocoa and Text Services Framework use UTF-16 for passing characters and
character indices while our app uses UTF-8 internally, so conversion
between them is necessary. One of things involved in the conversion is
selection ranges. They are measured from the start of a context string
(usually a paragraph or the whole text in a text field), so the
conversion can take a time proportional to the context string's length,
independent from the length of the portion the input system is actually
interested in.

This calls for an efficient algorithm. A rope structure has a good time
complexity and we already have one implementation, but it's complex and
overkill for this purpose when most of the context strings are expected
to be shorter than 1KB. Synchronizing the rope and the context string is
a hassle. Also, `Wm` currently doesn't have an API for partial updates
of a context string, diminishing the benefit of using such acceleration
structures.

The optimized linear-time algorithm implemented in `utf16count`
leverages a SIMD instruction set and some tricks to process multiple
bytes in a single iteration, providing a speed up by more than an order
of magnitude compared to a naïve implementation. Despite that, the code
size overhead is comparable to a naïve implementation.