feat(windows): implement the text input context API (WIP)
It's implemented using Text Services Framework (TSF) [1], the latest
text input framework which was first shipped with Office XP and has been
installed by default since Windows XP [2].

TSF is a successor to Input Method Manager (IMM). However, because of
the increased requirements on the applications, the adoption of TSF
is somewhat limited compared to IMM. For instance, GTK on Windows
doesn't use TSF at the point of writing. The attempt to support TSF in
Chrome was put on hold until it was completed [3] by Microsoft in 2019
as they started putting efforts to make their own fork of Chromium
dubbed as the next-gen Edge browser.

As of Windows 8, third-party IMEs are required to be TSF-aware [4], so
we can be assured that the users have a compatible IME.

[1]: https://docs.microsoft.com/en-us/windows/win32/tsf/text-services-framework
[2]: https://nyaruru.hatenablog.com/entry/20070309/p1
[3]: https://bugs.chromium.org/p/chromium/issues/detail?id=657623#c21
[4]: https://docs.microsoft.com/en-us/windows/win32/w8cookbook/third-party-input-method-editors