fix(macos): convert UTF-16 indices to UTF-8 in a pre-pass, resolving complexity requirement violation
In the previous implementation of `run_metrics_of_range`, the conversion
from UTF-16 indices to UTF-8 was done by caching the previous conversion
result and only recalculating the delta. With this approach,
`run_metrics_of_range` actually had the worst-case complexity of
`O(line_len²)` (where `line_len` represents the length of the specified
line). This was unacceptable and was a violation of the time complexity
requirement specified by the documentation of
`TextLayout::run_metrics_of_range`.

This commit solves this problem by incorporating the algorithm used in
the `windows` backend (added in 36a6e33).