perf(minisort): ensure the comparer is zero-sized whenever possible
libc's `qsort` function doesn't support passing context data separately,
so we use a global variable to pass one. The global variable has to be
protected from concurrent access by a mutex. Therefore it's important to
make the context data zero-sized whenever possible so that this overhead
can be elided.

All functions of the `minisort::qsort*` function family delegate the
processing to the core function `qsort_raw`. When doing this, they
provide a comparer function (which is the "context data"). The comparer
function converts the raw parameters for consumption by a user-provided
comparer function (which comes in multiple flavors), calls that
function, and converts the result back to the format the `libc::qsort`
function expects. Previously, the user-provided comparer function was
captured by reference, which accidentally prevented the context data
from being zero-sized.

> Moving `f` into the closure ensures the closure is zero-sized when `f`
> is zero-sized. The closure will never be zero-sized if `f was captured
> by reference.