perf(ui): remove the extra variant `Role::__Count`, eliminating bound checks
`tcw3::ui::theming::view::Shared` includes a field of type `[_;
ROLE_COUNT]`, which is logically equivalent to `HashMap<Role, _>`. The
array is indexed by the discriminant values of `Role`, but having the
extra variant `__Count` (which has the value `ROLE_COUNT`) meant bound
checks for this array couldn't be eliminated. This variant is used to
calculate the value of `ROLE_COUNT`.

This commit removes this extra variant, and instead derives the value of
`ROLE_COUNT` using the derive macro `variant_count::VariantCount`.