refactor(theming): put all styling rules' props into one `Vec`
`StylesheetMacroOutput` (the product of `stylesheet!`) includes a list
of prop values. Previously, there was a separate `Vec` storing prop
values for each styling rule. This commit changes this to use only one
`Vec` per `StylesheetMacroOutput`, reducing the number of dynamic
allocations.

This change actually increases the compiled executable size, which could
be because the compiler generates code that assigns the prop values
one by one, and the further they are away from the start, the larger
displacement values they need to access, the larger code each of them
needs to initialize.