feat(cggeom): redefine `box2!` to be `const fn`-compatible
It previously used `Into::into` to convert given values into `Point2` or
`Vector2`. In many cases, this allowed the endpoints to be specified in
a concise form like `[x, y]`, but had prevented the macro from being
used in a constant expression.

After this commit, the macro will not automatically use `Into::into`
anymore. This means the aforementioned concise form would be now a type
error. To make up for this, the macro will recognize inputs in the form
of `[x, y]` at a syntactic level and convert them into a desired type in
a `const fn`-compatible way. Owing to this mitigation, most of the
occurrences of `box2` didn't have to be rewritten.