perf(theming): split the styling prop initialization into a static part and a runtime part
The macro `stylesheet!` constructs `[(Prop, PropValue)]` containing all
prop values for all specified styling rules. This commit splits the
construction into a static part and a runtime part.

The first part is compiled into a statically-initialized array, which
the initialization code simply `memcpy`s to a dynamically-allocated
array. After that, the initialization code initializes the second part
one by one.

Most of the properties belong to the first part, so this change results
in a moderate code size reduction. The code size of the closure inside
`stella2::stylesheet::register_stylesheet` was cut down by more than
83%. The compiled executable shrank by 8KiB.