refactor(theming): rework the interface of `prop!`
Previously, `prop!(@value ...)` was used both for producing a generic
`PropValue` and for producing a constant `PropValue` to use during the
first pass. We could say the second one is the canonical usage and the
implementation of `port!(@dynvalue ...)` (sort of) abused it to cover
the first use case.

The reworked interface adds a input pattern `prop!(@constvalue ...)`,
which is dedicated for producing a `PropValue` for the first pass.

Also, `@dynvalue` was renamed to `@setdynvalue` for clarification.