perf(theming): box the inner value of some variants of `PropValue`
This commit reduces the size of `PropValue` (from 32 to 24 on 64-bit
platforms) by boxing some of its variants, yielding code size reduction
and working set reduction. The code size reduction is attributed mainly
to the following factors: (1) the more optimal code generation
(specifically, the removal of unnecessary register-to-stack-to-array
copies) of `new_custom_stylesheet`, which may or may not be directly
connected with this change, and (2) the sheer smaller size of the static
prop values.

The boxing solution implemented by this change doesn't use `Box<T>` for
boxing, for `Box<T>` cannot be constructed during a compile-time
evaluation. Instead, we wrap the values with `rob::Rob<'a, T>`, which is
essentially a tagged union having variants `&'a T` and `Box<T>`,
paralleling the `Cow` from the standard library.