chore(icon_baker): remove dependency on `image`'s default features
This commit disables the default feature (viz. `image`) of `nsvg`. This
only removes the method `SvgImage::rasterize`, which is practically a
wrapper around the lower-level method `rasterize_to_raw_rgba`

`icon_baker` still needs `image` for other things, so it now depends
directly on `image`, but with all default features disabled. This means
we now have a direct control on `image`'s features means and we can omit
a great deal of `image`'s unused features.