fix(view): do not include the search box in the `StyledBox` having a `FOCUS` auto class
The channel list items are rendered with a "focused" style when there's
an ancestor having a `FOCUS` class. One of their ancestors is a
`StyledBox` having `FOCUS` in its `auto_class_set`, which also included
the search box. As a result, the items were erroneously rendered as
focused when it's the search box that actually got a keyboard focus.

This commit rearranges the view hierarchy to address this issue.