fix(theming): use direction-specific `layer_center` values for scrollbar elements
`layer_center` is used to specify the thickness of each edge in 9-slice
scaling. There are inconsistencies between platforms regarding how to
handle the cases in which the opposing edges overlap with each other. Of
all supported platforms, only Windows exhibits an unintuitive behavior
where edges are crushed non-uniformly. This causes the scrollbar
elements to look asymmetrical and deformed.

This commit fixes this problem by setting `layer_center` in a way that
doesn't cause overlapping edges.