feat(theming): add `Prop::AllowGrow`
The `AllowGrow` prop has a type `[bool; 2]`, each flag specifying
whether the element can expand in the corresponding axis or not.

Previously, the only mechanism that could be used to limit the size of
a `StyledBox` was to use `Prop::SubviewMetrics`. Not only was this a
hacky solution, it was also only applicable when it had a subview. The
addition of `AllowGrow` removes this limitation.