fix(macos): override `respondsToSelector:` to dynamically define methods
The `macos` backend adds methods to an Objective C class at runtime so
that the set of handled selectors can be dynamically specified by the
accelerator tables given by the application.

The old method involves the use of `[NSObject resolveInstanceMethod:]`,
which is the documented way to resolve methods at runtime. The problem
is, the runtime apparently caches the result of the method, so once it
returns `NO`, it won't be called again, and accelerator tables activated
later might not take effect.

This commit solves this problem by overriding `[NSObject
respondsToSelector:]` instead. Cocoa (presumably) calls this method
while traversing through a responder chain, so the runtime will not have
a chance to cache the result.