feat(pal): define text input actions
`NSTextInputClient` (Cocoa), in addition to text insertion, also handles
miscellaneous input commands such as "move to right" and generates Obj-C
messages defined in `NSStandardKeyBindingResponding`. This commit
defines how to handle such commands in TCW3. This is implementable by
other backends as well by hard-coding key bindings in the backend.

The key bindings for `NSStandardKeyBindingResponding` are completely
re-mappable by the user [1][2], which is why we should rely on this
mechanism.

One possible alternative is to add a new method to
`TextInputCtxListener` for handling text input actions. I decided not to
take this route because there are some overlaps between
`NSStandardKeyBindingResponding` and common actions (e.g., `SELECT_ALL`)
that may be used in other places, and it would be unreasonable for
widgets to handle both pathways of delivery for the same action so that
it works regardless of whether the application defines a shadowing key
binding for the said action.

[1]: https://developer.apple.com/library/archive/documentation/Cocoa/Conceptual/EventOverview/TextDefaultsBindings/TextDefaultsBindings.html
[2]: https://apple.stackexchange.com/questions/16135/remap-home-and-end-to-beginning-and-end-of-line