fix(pal): note that`SELECT_ALL` isn't implicitly translated
Although `selectAll:` is in `NSStandardKeyBindingResponding`, Cocoa
doesn't translate "Command + A" into this. I suspect this is because an
application main menu usually contains "Select All" with "Command + A"
being its key equivalent, and the user would be confused if the original
mapping is still working even after changing its key equivalent.