fix(macos): clear the composition range when it's empty
In some cases, `setMarkedText:...` empties the composition range by
removing all characters from it. Previously, this code did not clear the
range (i.e., assign `None`) in such a case. However, I found that
Kotoeri misbehaves with this implementation.

Kotoeri (macOS's standard Japanese IME), when the cursor is at the start
of the composition, allows the user to include a preceding text into the
current composition simply by pressing the left arrow key. Before this
commit, when the user did this, the preceding text was transfered to the
composition, but it wasn't removed from the original position, causing
it to be duplicated every time the user hit the left arrow key.

This commit fixes this problem.