feat(uicore): add `ViewFlags::STRONG_FOCUS`
This commit partly imitates the focus behavior of macOS. The widgets
requiring a keyboard focus for normal interaction should have the flag
`STRONG_FOCUS`. The focus-on-click behavior is restricted to the widgets
with this flag, meaning widgets will not receive a focus by clicking
unless it's absolutely necessary.

Tab navigation is unaffected by this flag. macOS use a different
behavior by default, but this can be changed by the user. We could
detect the current user preference, but there's no obvious way to do so,
so we play it safe to prevent the application from being completely
unusable without a mouse.