feat(uicore): make `set_focused_view` search children for a focusable view
Some widgets such as `Slider`¹ have a `StyledBox` with `TAB_STOP` to
reflect the focus state to the appearance, and wrap it with a bare view
with `ACCEPT_MOUSE_DRAG` to receive mouse events. With this
configuration, when the widget is clicked, the system will only checks
the outer view's ancestors and won't be able to find that the inner view
is focusable.

This commit modifies `set_focused_view` to search subviews as well.

¹ Actually, `Slider` doesn't have `STRONG_FOCUS`, so it won't receive a
focus by clicking. However, there are no other widgets having a similar
structure which is suitable as an example. Thus, we assume that `Slider`
has `STRONG_FOCUS` in this discourse.