fix(windows): call `key_down` if `validate_action` doesn't return `VALID`
Fixes tab navigation not working in an `Entry`.

The old behavior was to drop key events if they were translated by
accelerator tables but the current `WndListener` does not recognize the
translated actions. This meant certain keys were consistently
interpreted as actions if they are in an accelerator table, but that
also meant certain keys hard-coded in the text input accelerator table
were never seen by `key_down`.

This commit changes this to drop key events only if the flag returned by
`validate_action` includes `VALID` but does not include `ENABLED`.