fix(unix): fix the struct layout of `PangoLogAttr`
gtk-rs/gir (automatic binding generator for gtk-rs) currently don't
support[1] structs with bitfields and translates them to
incorrectly-sized `struct`s. As a result, `pango_sys::PangoLogAttr` was
twice as long as it was supposed to be, and the output of
`get_log_attrs_readonly` referred out of bounds. This is why we saw a
text cursor in an `Entry` skipping characters when the user pressed
arrow keys.

This commit addresses this problem by interpreting `PangoLogAttr*` as
`LogAttrFlags*` (which we already have) and carefully laying out
`LogAttrFlags` according to `PangoLogAttr`.

[1]: https://github.com/gtk-rs/gir/issues/465