feat(uicore): add `HView::global_visible_frame`
`global_visible_frame` represents the bounding rectangle of the visible
portion of a view.

This will be useful when you have a view much larger than a viewport,
only a small portion of which is visible at once. `CanvasMixin` renders
an entire view for now, which doesn't scale well for large views in
terms of memory consumption. With this addition, it could be modified to
divide the view into tiles and paint individual tiles only as needed.