refactor(pal): add `Init` and `SendInit` to facilitate the safe initialization of `MtSticky`
This commit adds traits `Init` and `SendInit` to allow the static
initialization of variables of type `MtSticky` containing unsendable
collection types in a safe manner (the default values of such collection
types are safe to send because they don't actually contain anything
unsendable).

In other words, the code comments saying "(type name) is `!Send`, but
there is no instance at this point, so `new_unchecked` is safe" are
now replaced with a safe, proven solution.